Refactor query/ in preparation for new commands. Factor out EXPLAIN logic from mysql/mysql.go to query/mysql/executor.go. Factor out AtLeastVersion logic from mysql/mysql.go to pct/sys.go. Make query.Manager stoppable. query/service/ is now query/mysql/.
