+++ query/manager.go
@@ -18,9 +18,15 @@
 package query
 
 import (
-	"github.com/percona/cloud-protocol/proto"
-	"github.com/percona/percona-agent/pct"
+	"encoding/json"
+	"fmt"
 	"sync"
+
+	"github.com/percona/cloud-protocol/proto"
+	"github.com/percona/percona-agent/instance"
+	"github.com/percona/percona-agent/mysql"
+	"github.com/percona/percona-agent/pct"
+	mysqlExec "github.com/percona/percona-agent/query/mysql"
 )
 
 const (
@@ -28,19 +34,20 @@ const (
 )
 
 type Manager struct {
-	logger  *pct.Logger
-	explain Service
+	logger       *pct.Logger
+	instanceRepo *instance.Repo
+	connFactory  mysql.ConnectionFactory
 	// --
 	running bool
 	sync.Mutex
-	// --
 	status *pct.Status
 }
 
-func NewManager(logger *pct.Logger, explain Service) *Manager {
+func NewManager(logger *pct.Logger, instanceRepo *instance.Repo, connFactory mysql.ConnectionFactory) *Manager {
 	m := &Manager{
-		logger:  logger,
-		explain: explain,
+		logger:       logger,
+		instanceRepo: instanceRepo,
+		connFactory:  connFactory,
 		// --
 		status: pct.NewStatus([]string{SERVICE_NAME}),
 	}
@@ -54,19 +61,25 @@ func NewManager(logger *pct.Logger, explain Service) *Manager {
 func (m *Manager) Start() error {
 	m.Lock()
 	defer m.Unlock()
-
 	if m.running {
 		return pct.ServiceIsRunningError{Service: SERVICE_NAME}
 	}
-
 	m.running = true
 	m.logger.Info("Started")
-	m.status.Update(SERVICE_NAME, "Running")
+	m.status.Update(SERVICE_NAME, "Idle")
 	return nil
 }
 
 func (m *Manager) Stop() error {
-	// Can't stop this manager.
+	// Let user stop this tool in case they don't want agent executing queries.
+	m.Lock()
+	defer m.Unlock()
+	if !m.running {
+		return nil
+	}
+	m.running = false
+	m.logger.Info("Stopped")
+	m.status.Update(SERVICE_NAME, "Stopped")
 	return nil
 }
 
@@ -74,15 +87,26 @@ func (m *Manager) Handle(cmd *proto.Cmd) *proto.Reply {
 	m.Lock()
 	defer m.Unlock()
 
-	m.status.UpdateRe(SERVICE_NAME, "Handling", cmd)
-	defer m.status.Update(SERVICE_NAME, "Running")
+	// Don't query if this tool is stopped.
+	if !m.running {
+		return cmd.Reply(nil, pct.ServiceIsNotRunningError{})
+	}
 
-	switch cmd.Cmd {
-	case "Explain":
-		m.status.UpdateRe(SERVICE_NAME, "Running explain", cmd)
-		return m.explain.Handle(cmd)
+	m.status.UpdateRe(SERVICE_NAME, "Handling", cmd)
+	defer m.status.Update(SERVICE_NAME, "Idle")
+
+	// See which type of subsystem this query is for. Right now we only support
+	// MySQL, but this abstraction will make adding other subsystems easy.
+	si := &proto.ServiceInstance{}
+	if err := json.Unmarshal(cmd.Data, si); err != nil {
+		return cmd.Reply(nil, err)
+	}
+
+	switch si.Service {
+	case "mysql":
+		return m.handleMySQLQuery(cmd, si)
 	default:
-		return cmd.Reply(nil, pct.UnknownCmdError{Cmd: cmd.Cmd})
+		return cmd.Reply(nil, pct.UnknownCmdError{si.Service})
 	}
 }
 
@@ -92,4 +116,40 @@ func (m *Manager) Status() map[string]string {
 
 func (m *Manager) GetConfig() ([]proto.AgentConfig, []error) {
 	return nil, nil
+}
+
+// --------------------------------------------------------------------------
+
+func (m *Manager) handleMySQLQuery(cmd *proto.Cmd, si *proto.ServiceInstance) *proto.Reply {
+	// Connect to MySQL.
+	mysqlIt := &proto.MySQLInstance{}
+	if err := m.instanceRepo.Get(si.Service, si.InstanceId, mysqlIt); err != nil {
+		return cmd.Reply(nil, err)
+	}
+	conn := m.connFactory.Make(mysqlIt.DSN)
+	if err := conn.Connect(1); err != nil {
+		return cmd.Reply(nil, fmt.Errorf("Cannot connect to MySQL: %s", err))
+	}
+	defer conn.Close()
+
+	// Get the instance name, e.g. mysql-db01, to make status human-readable.
+	instanceName := m.instanceRepo.Name(si.Service, si.InstanceId)
+
+	// Execute the query.
+	switch cmd.Cmd {
+	case "Explain":
+		m.status.Update(SERVICE_NAME, "EXPLAIN query on "+instanceName)
+		q := &proto.ExplainQuery{}
+		if err := json.Unmarshal(cmd.Data, q); err != nil {
+			return cmd.Reply(nil, err)
+		}
+		e := mysqlExec.NewQueryExecutor(conn)
+		res, err := e.Explain(q.Db, q.Query)
+		if err != nil {
+			return cmd.Reply(nil, fmt.Errorf("EXPLAIN failed: %s", err))
+		}
+		return cmd.Reply(res, nil)
+	default:
+		return cmd.Reply(nil, pct.UnknownCmdError{Cmd: cmd.Cmd})
+	}
 }
