+++ query/query_test.go
@@ -18,16 +18,22 @@
 package query_test
 
 import (
+	"encoding/json"
 	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"testing"
+
 	"github.com/percona/cloud-protocol/proto"
+	"github.com/percona/percona-agent/instance"
+	"github.com/percona/percona-agent/mysql"
 	"github.com/percona/percona-agent/pct"
 	"github.com/percona/percona-agent/query"
 	"github.com/percona/percona-agent/test/mock"
 	. "gopkg.in/check.v1"
-	"testing"
 )
 
-// Hook up gocheck into the "go test" runner.
 func Test(t *testing.T) { TestingT(t) }
 
 /////////////////////////////////////////////////////////////////////////////
@@ -35,34 +41,80 @@ func Test(t *testing.T) { TestingT(t) }
 /////////////////////////////////////////////////////////////////////////////
 
 type ManagerTestSuite struct {
-	logChan chan *proto.LogEntry
-	logger  *pct.Logger
+	logChan       chan *proto.LogEntry
+	logger        *pct.Logger
+	configDir     string
+	tmpDir        string
+	dsn           string
+	repo          *instance.Repo
+	mysqlInstance proto.ServiceInstance
+	api           *mock.API
 }
 
 var _ = Suite(&ManagerTestSuite{})
 
 func (s *ManagerTestSuite) SetUpSuite(t *C) {
+	s.dsn = os.Getenv("PCT_TEST_MYSQL_DSN")
+	if s.dsn == "" {
+		t.Fatal("PCT_TEST_MYSQL_DSN is not set")
+	}
+
 	s.logChan = make(chan *proto.LogEntry, 10)
 	s.logger = pct.NewLogger(s.logChan, query.SERVICE_NAME+"-manager-test")
+
+	var err error
+	s.tmpDir, err = ioutil.TempDir("/tmp", "agent-test")
+	t.Assert(err, IsNil)
+
+	if err := pct.Basedir.Init(s.tmpDir); err != nil {
+		t.Fatal(err)
+	}
+	s.configDir = pct.Basedir.Dir("config")
+
+	// Real instance repo
+	s.repo = instance.NewRepo(pct.NewLogger(s.logChan, "im-test"), s.configDir, s.api)
+	data, err := json.Marshal(&proto.MySQLInstance{
+		Hostname: "db1",
+		DSN:      s.dsn,
+	})
+	t.Assert(err, IsNil)
+	s.repo.Add("mysql", 1, data, false)
+	s.mysqlInstance = proto.ServiceInstance{Service: "mysql", InstanceId: 1}
+
+	links := map[string]string{
+		"agent":     "http://localhost/agent",
+		"instances": "http://localhost/instances",
+	}
+	s.api = mock.NewAPI("http://localhost", "http://localhost", "123", "abc-123-def", links)
 }
 
 func (s *ManagerTestSuite) SetUpTest(t *C) {
+	glob := filepath.Join(pct.Basedir.Dir("config"), "*")
+	files, err := filepath.Glob(glob)
+	if err != nil {
+		t.Fatal(err)
+	}
+	for _, file := range files {
+		if err := os.Remove(file); err != nil {
+			t.Fatal(err)
+		}
+	}
 }
 
 func (s *ManagerTestSuite) TearDownSuite(t *C) {
+	if err := os.RemoveAll(s.tmpDir); err != nil {
+		t.Error(err)
+	}
 }
 
 // --------------------------------------------------------------------------
 
-func (s *ManagerTestSuite) TestStartStopHandleManager(t *C) {
+func (s *ManagerTestSuite) TestStartStop(t *C) {
 	var err error
 
-	// Create explain service
-	explainService := mock.NewQueryService()
-
 	// Create query manager
-	m := query.NewManager(s.logger, explainService)
-	t.Assert(m, Not(IsNil), Commentf("Make new query.Manager"))
+	m := query.NewManager(s.logger, s.repo, &mysql.RealConnectionFactory{})
+	t.Assert(m, NotNil)
 
 	// The agent calls mm.Start().
 	err = m.Start()
@@ -70,33 +122,63 @@ func (s *ManagerTestSuite) TestStartStopHandleManager(t *C) {
 
 	// Its status should be "Running".
 	status := m.Status()
-	t.Check(status[query.SERVICE_NAME], Equals, "Running")
+	t.Check(status[query.SERVICE_NAME], Equals, "Idle")
 
 	// Can't start manager twice.
 	err = m.Start()
 	t.Check(err, FitsTypeOf, pct.ServiceIsRunningError{})
 
+	// Stop the service.
+	err = m.Stop()
+	t.Check(err, IsNil)
+	status = m.Status()
+	t.Check(status[query.SERVICE_NAME], Equals, "Stopped")
+
+	// Stop is idempotent.
+	err = m.Stop()
+	t.Check(err, IsNil)
+	status = m.Status()
+	t.Check(status[query.SERVICE_NAME], Equals, "Stopped")
+}
+
+func (s *ManagerTestSuite) TestHandleExplain(t *C) {
+	var err error
+
+	// Create query manager
+	m := query.NewManager(s.logger, s.repo, &mysql.RealConnectionFactory{})
+	t.Assert(m, NotNil)
+
+	// The agent calls mm.Start().
+	err = m.Start()
+	t.Assert(err, IsNil)
+
 	// Test known cmd
+	query := proto.ExplainQuery{
+		ServiceInstance: proto.ServiceInstance{
+			Service:    "mysql",
+			InstanceId: 1,
+		},
+		Query: "SELECT 1",
+	}
+	data, err := json.Marshal(query)
+	t.Assert(err, IsNil)
 	cmd := &proto.Cmd{
 		Service: "query",
 		Cmd:     "Explain",
+		Data:    data,
 	}
 	gotReply := m.Handle(cmd)
 	t.Assert(gotReply, NotNil)
-	t.Assert(gotReply.Error, Equals, "")
+	t.Check(gotReply.Error, Equals, "")
+	t.Check(gotReply.Data, Not(HasLen), 0)
 
 	// Test unknown cmd
 	cmd = &proto.Cmd{
 		Service: "query",
 		Cmd:     "Unknown",
+		Data:    data,
 	}
 	gotReply = m.Handle(cmd)
 	t.Assert(gotReply, NotNil)
 	t.Assert(gotReply.Error, Equals, fmt.Sprintf("Unknown command: %s", cmd.Cmd))
-
-	// You can't stop this service
-	err = m.Stop()
-	t.Check(err, IsNil)
-	status = m.Status()
-	t.Check(status[query.SERVICE_NAME], Equals, "Running")
 }
